Organize and hover color on blog item
