+++ src/lib/components/TagList.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Tag from '$lib/components/Tag.svelte'
+	import Tag from '$lib/components/elements/Tag.svelte'
 
 	export let tags
 </script>
