+++ src/lib/components/blogs/BlogItem.svelte
@@ -6,17 +6,18 @@
 </script>
 
 <article>
-	<div class="flex flex-col gap-3 rounded-xl bg-arkGray5 p-5">
+	<a
+		href="{base}/blog/{post.slug}"
+		class="flex flex-col gap-3 rounded-xl bg-arkGray5 p-5 hover:bg-[#FCF0E3]"
+	>
 		<div class="flex flex-col">
-			<a href="{base}/blog/{post.slug}" class="text-2xl font-extrabold hover:underline"
-				>{post.title}</a
-			>
+			<p class="text-2xl font-extrabold">{post.title}</p>
 			<p class="font-medium text-arkGray4">{post.date}</p>
 		</div>
 
 		<div class="h-full max-h-36 overflow-hidden">
 			{post.summary}
 		</div>
-		<a href="{base}/blog/{post.slug}" class="w-fit font-bold text-arkOrange underline">Read</a>
-	</div>
+		<a href="{base}/blog/{post.slug}" class="w-fit font-bold text-arkOrange">Read</a>
+	</a>
 </article>
