+++ src/routes/blog/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
+	import Blogs from '$lib/components/blogs/Blogs.svelte'
 	import Head from '$lib/components/layouts/Head.svelte'
-	import Blogs from '$lib/components/Blogs.svelte'
 
 	export let data
 	const { posts, tags } = data
