+++ src/routes/blog/[slug]/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Blog from '$lib/components/Blog.svelte'
+	import Blog from '$lib/components/blogs/Blog.svelte'
 	import Head from '$lib/components/layouts/Head.svelte'
 
 	export let data
